implemented tests for webserver
