+++ savant_core/src/webserver/kvs_handlers.rs
@@ -81,185 +81,3 @@ async fn get_handler(path: web::Path<(String, String)>) -> HttpResponse {
         HttpResponse::InternalServerError().finish()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::get_or_init_async_runtime;
-    use crate::primitives::attribute_set::AttributeSet;
-    use crate::primitives::Attribute;
-    use crate::protobuf::{from_pb, ToProtobuf};
-    use crate::webserver::kvs::synchronous::get_attribute;
-    use crate::webserver::kvs::synchronous::set_attributes;
-    use crate::webserver::{init_webserver, set_status, stop_webserver, PipelineStatus};
-    use savant_protobuf::generated;
-    use std::thread::sleep;
-    use std::time::Duration;
-
-    #[test]
-    #[serial_test::serial]
-    fn test_attributes_abi_to_api() -> anyhow::Result<()> {
-        init_webserver(8888)?;
-        sleep(Duration::from_millis(100));
-        set_status(PipelineStatus::Running)?;
-        let ttl_attribute_set = AttributeSet::from(vec![Attribute::persistent(
-            "jkl",
-            "yay",
-            vec![],
-            &None,
-            false,
-        )]);
-        let attribute_set = AttributeSet::from(vec![Attribute::persistent(
-            "ghi",
-            "yay",
-            vec![],
-            &None,
-            false,
-        )]);
-        set_attributes(&ttl_attribute_set, Some(500));
-        set_attributes(&attribute_set, None);
-
-        let r = reqwest::blocking::get("http://localhost:8888/kvs/search-keys/*/*")?;
-        assert_eq!(r.status(), 200);
-        let mut result: Vec<(String, String)> = r.json()?;
-        result.sort_by(|(a, _), (b, _)| a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal));
-        assert_eq!(
-            result,
-            vec![
-                ("ghi".to_string(), "yay".to_string()),
-                ("jkl".to_string(), "yay".to_string())
-            ]
-        );
-        sleep(Duration::from_millis(501));
-
-        let r = reqwest::blocking::get("http://localhost:8888/kvs/search-keys/*/*")?;
-        assert_eq!(r.status(), 200);
-        let result: Vec<(String, String)> = r.json()?;
-        assert_eq!(result, vec![("ghi".to_string(), "yay".to_string())]);
-
-        let r = reqwest::blocking::get("http://localhost:8888/kvs/search/*/*")?;
-        assert_eq!(r.status(), 200);
-        let binary = r.bytes()?;
-        let res_attribute_set = from_pb::<generated::AttributeSet, AttributeSet>(&binary)?;
-        assert_eq!(res_attribute_set.attributes, attribute_set.attributes);
-
-        let r = reqwest::blocking::get("http://localhost:8888/kvs/get/ghi/yay")?;
-        assert_eq!(r.status(), 200);
-        let binary = r.bytes()?;
-        let res_attribute = from_pb::<generated::Attribute, Attribute>(&binary)?;
-        assert_eq!(res_attribute, attribute_set.attributes[0]);
-
-        let rt = get_or_init_async_runtime();
-        let client = reqwest::Client::new();
-        // delete single
-        let r = rt.block_on(async {
-            let resp = client
-                .post("http://localhost:8888/kvs/delete-single/ghi/yay")
-                .send()
-                .await
-                .unwrap();
-            assert_eq!(resp.status(), 200);
-            resp.bytes().await
-        })?;
-        let res_attribute = from_pb::<generated::Attribute, Attribute>(&r)?;
-        assert_eq!(res_attribute, attribute_set.attributes[0]);
-
-        // delete after delete
-        let r = rt.block_on(async {
-            let resp = client
-                .post("http://localhost:8888/kvs/delete-single/ghi/yay")
-                .send()
-                .await
-                .unwrap();
-            assert_eq!(resp.status(), 200);
-            resp.bytes().await
-        })?;
-        assert_eq!(r.len(), 0);
-
-        // set again and purge
-        set_attributes(&attribute_set, None);
-        let r = rt.block_on(async {
-            let resp = client
-                .post("http://localhost:8888/kvs/purge/*/yay")
-                .send()
-                .await
-                .unwrap();
-            assert_eq!(resp.status(), 200);
-            resp.bytes().await
-        })?;
-        assert_eq!(r.len(), 0); // returns nothing
-                                // ensure that nothing exists with get
-        let r = rt.block_on(async {
-            let resp = client
-                .get("http://localhost:8888/kvs/get/ghi/yay")
-                .send()
-                .await
-                .unwrap();
-            assert_eq!(resp.status(), 200);
-            resp.bytes().await
-        })?;
-        assert_eq!(r.len(), 0);
-
-        stop_webserver();
-        Ok(())
-    }
-
-    #[test]
-    #[serial_test::serial]
-    fn test_api_to_abi() -> anyhow::Result<()> {
-        init_webserver(8888)?;
-        sleep(Duration::from_millis(100));
-        set_status(PipelineStatus::Running)?;
-        let ttl_attribute_set = AttributeSet::from(vec![Attribute::persistent(
-            "jkl",
-            "yay",
-            vec![],
-            &None,
-            false,
-        )]);
-        let attribute_set = AttributeSet::from(vec![Attribute::persistent(
-            "ghi",
-            "yay",
-            vec![],
-            &None,
-            false,
-        )]);
-
-        // set without ttl
-        let rt = get_or_init_async_runtime();
-        let client = reqwest::Client::new();
-
-        let r = rt.block_on(async {
-            let resp = client
-                .post("http://localhost:8888/kvs/set")
-                .body(attribute_set.to_pb().unwrap())
-                .send()
-                .await
-                .unwrap();
-            assert_eq!(resp.status(), 200);
-            resp.bytes().await
-        })?;
-        assert_eq!(r.len(), 0);
-        let attr = get_attribute(&"ghi".to_string(), &"yay".to_string());
-        assert_eq!(attr.unwrap(), attribute_set.attributes[0]);
-
-        // set with ttl
-        let _ = rt.block_on(async {
-            let resp = client
-                .post("http://localhost:8888/kvs/set-with-ttl/500")
-                .body(ttl_attribute_set.to_pb().unwrap())
-                .send()
-                .await
-                .unwrap();
-            assert_eq!(resp.status(), 200);
-            resp.bytes().await
-        })?;
-        let attr = get_attribute(&"jkl".to_string(), &"yay".to_string());
-        assert_eq!(attr.unwrap(), ttl_attribute_set.attributes[0]);
-        sleep(Duration::from_millis(501));
-        let attr = get_attribute(&"jkl".to_string(), &"yay".to_string());
-        assert!(attr.is_none());
-
-        stop_webserver();
-        Ok(())
-    }
-}
